dragonwind_simulator/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Data processing and analysis
xarray>=0.19.0
netCDF4>=1.5.7
h5py>=3.6.0
pyarrow>=6.0.0

# Machine learning and optimization
scikit-learn>=1.0.0
tensorflow>=2.8.0  # or pytorch>=1.10.0
optuna>=2.10.0

# Geospatial analysis
geopandas>=0.10.0
rasterio>=1.2.0
shapely>=1.8.0

# Visualization
plotly>=5.3.0
dash>=2.0.0
bokeh>=2.4.0

# Testing and development
pytest>=6.2.5
pytest-cov>=2.12.0
black>=21.12b0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.910

# Documentation
sphinx>=4.2.0
sphinx-rtd-theme>=1.0.0
nbsphinx>=0.8.8

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.4.0

# Other utilities
tqdm>=4.62.0
python-dotenv>=0.19.0
pyyaml>=6.0.0
